requirements.txt
python-dotenv
flask
flask-login
oauthlib
flask-session
pyyaml
redis
requests
Flask_SocketIO
